refactor targeted fuzz into separate tests (#16879)
## Description

This PR refactors the `targetedFuzz.spec.ts` file into separate test
files
